requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anthropic==0.3.10
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
av==10.0.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
bokeh==2.4.1
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
distro==1.8.0
docopt==0.6.2
duckduckgo-search==3.8.5
exceptiongroup==1.1.3
Faker==19.3.0
fastapi==0.101.1
favicon==0.7.0
ffmpeg==1.4
ffmpy==0.3.1
filelock==3.12.2
Flask==2.3.2
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
gevent==1.4.0
gitdb==4.0.10
GitPython==3.1.32
gradio==3.40.1
gradio_client==0.4.0
greenlet==0.4.15
gTTS==2.3.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htbuilder==0.6.1
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.74
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
langchain==0.0.268
langsmith==0.0.25
linkify-it-py==2.0.2
loguru==0.7.0
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==10.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.24.4
openai==0.27.8
orjson==3.9.5
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pipwin==0.5.2
pkgutil_resolve_name==1.3.10
protobuf==4.24.1
pyarrow==12.0.1
PyAudio==0.2.11
pycparser==2.19
pydantic==1.10.12
pydantic_core==2.6.1
pydeck==0.8.0
pydub==0.25.1
Pygments==2.16.1
pyjsparser==2.7.1
pymdown-extensions==10.1
Pympler==1.0.1
pyparsing==3.0.9
PyPrind==2.11.3
pyproject==1.3.1
pySmartDL==1.3.4
python-dateutil==2.8.2
python-docx==0.8.11
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
semantic-version==2.10.0
six==1.12.0
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
sounddevice==0.4.6
soupsieve==2.4.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.20
st-annotated-text==4.0.0
starlette==0.27.0
streamlit==1.25.0
streamlit-bokeh-events==0.1.2
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-chat==0.1.1
streamlit-embedcode==0.1.2
streamlit-extras==0.3.0
streamlit-faker==0.0.2
streamlit-feedback==0.0.9
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.3
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
trubrics==1.4.5
typeguard==2.13.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
urllib3==2.0.4
uvicorn==0.23.2
validators==0.21.2
watchdog==3.0.0
Wave==0.0.2
websocket==0.2.1
websocket-client==0.58.0
websockets==11.0.3
Werkzeug==2.3.7
xyzservices==2023.7.0
yarl==1.9.2
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
